Modifico la tabla de viajes para que deje de usar decimales en inicio de caja y en gastos

+++ main/models.py
@@ -168,10 +168,10 @@ class Viaje(models.Model):
     # TODO: LLAVES FORANEAS ¿Remitos?
     chofer = models.ForeignKey(Chofer, on_delete=models.PROTECT, db_column="id_chofer")
     vehiculo = models.ForeignKey(Vehiculo, on_delete=models.PROTECT, db_column="id_vehiculo")
-    inicio_caja = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    inicio_caja = models.IntegerField(default=0)
     fecha_inicio = models.DateField()
     fecha_vuelta = models.DateField(null=True, blank=True)
-    gastos = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    gastos = models.IntegerField(default=0)
     activo = models.BooleanField(default=True)
 
     @property
